move new user methode to user helper
