+++ app/controllers/users_controller.rb
@@ -23,15 +23,4 @@ class UsersController < ApplicationController
   def param_user
     params.require(:user).permit(:username, :avatar)
   end
-
-  def set_user
-    @user = User.find(params[:id])
-  end
-
-  def require_same_user
-    if current_user != @user
-      flash[:danger] = 'you do not have permission to peform this'
-      redirect_to root_path
-    end
-  end
 end
